+++ app/src/main/kotlin/org/fossify/musicplayer/interfaces/SongsDao.kt
@@ -44,6 +44,9 @@ interface SongsDao {
     @Query("DELETE FROM tracks WHERE playlist_id = :playlistId")
     fun removePlaylistSongs(playlistId: Int)
 
+    @Query("DELETE FROM tracks WHERE playlist_id = :playlistId AND media_store_id = :mediaStoreId")
+    fun removeTrackFromPlaylist(playlistId: Int, mediaStoreId: Long)
+
     @Query("UPDATE tracks SET path = :newPath, artist = :artist, title = :title WHERE path = :oldPath")
     fun updateSongInfo(newPath: String, artist: String, title: String, oldPath: String)